Delete orphaned re-fly anchor-selection bypass + force-Absolute toggle
The narrowed-gate filter (FilterCandidatesForReFlyProvisional) is the
validated default, so the one-release rollback path left by PR #909 is
now removed:

- ReFlyAnchorSelection: deleted both TryResolveReFlyProvisionalAnchor
  overloads + the private walk/resolver helpers and their constants;
  rewrote the class docstring to describe only the surviving filter +
  predicate members.
- FlightRecorder / BackgroundRecorder: deleted the two apply helpers
  (ApplyReFlyProvisionalAnchorToActiveRecording /
  ApplyReFlyProvisionalAnchorToState) and the force-Absolute gate blocks
  in UpdateAnchorDetection, UpdateBackgroundAnchorDetection, and
  RestoreTrackSectionAfterFalseAlarm; trimmed stale bypass comments.
- AnchorDetector: deleted the now-unreferenced
  AnchorCandidateSource.ReFlyProvisionalSupersede enum value (in-memory
  diagnostic only, never serialized).
- ParsekSettings / ParsekSettingsPersistence / SettingsWindowUI: deleted
  the forceAbsoluteForReFlyProvisional field, property, notify helper,
  persistence keys/store/load/save/reconcile, test shims, and UI toggle.

Tests: deleted ReFlyAnchorSelectionTests and the bypass/toggle in-game
tests; moved the IsActiveRecordingReFlyProvisional predicate tests into
FilterCandidatesForReFlyProvisionalTests (the filter depends on the
predicate); retargeted the wiring gate onto the filter call.

No schema bump: this removes a recorder code path + a setting, not a
.prec/.pann field. A missing forceAbsoluteForReFlyProvisional key loads
as the using-default branch.
