+++ CHANGELOG.md
@@ -57,8 +57,8 @@ All notable changes to Parsek are documented here.
 - Pre-switch dialog Round-6 review follow-up: context-aware ScreenMessage and ledger reason on no-session Discard, plus minor diagnostic log hygiene.
 - Watch auto-follow no longer logs a three-line WARN cascade while waiting for the chain-continuation ghost to spawn. The deferred-transfer branch in `WatchModeController.TransferWatchToNextSegment` now routes through a target-keyed rate-limited Verbose helper (`LogAutoFollowDeferred`), so a normal one- or two-frame wait emits a single Verbose line per chain transfer instead of three WARN lines.
 - Internal refactor: `GhostPlaybackEngine` no longer reads `SessionSuppressionState` directly; the active re-fly marker and per-recording suppression bit flow in via `FrameContext` / `TrajectoryPlaybackFlags`. No behavioral change.
-- Re-fly Relative anchor selection now uses a narrowed-gate filter instead of the PR #889 supersede-target bypass. While a re-fly session is active, the nearest-search drops every candidate whose recording is a member of the same `RecordingTree` as the provisional, leaving out-of-tree real anchors (stations, bases, live vessels from other lineages) eligible. Net result: re-fly with no nearby real anchor authors Absolute (the PR 901 validated default), re-fly mid-docking-approach authors Relative-against-real-station (closes the PR 901 "docking-mid-rewind" regression), and re-fly inside a loop chain with a live loop anchor authors Relative-against-live-loop-anchor (closes the "loop-anchored re-fly fork" regression). The `forceAbsoluteForReFlyProvisional` setting remains as a rollback path that forces fully Absolute even when a real anchor is nearby. `ReFlyAnchorSelection.TryResolveReFlyProvisionalAnchor` and the two `ApplyReFlyProvisionalAnchor*` apply helpers are orphaned (no callers) but retained for one release; scheduled for deletion in a follow-up PR.
-- New experimental setting `forceAbsoluteForReFlyProvisional` (off by default, under Settings > Diagnostics) to A/B compare the current Relative-against-superseded-origin authoring against the simpler debris-style Absolute path for re-fly provisional recordings. When on, the recorder skips both the `ReFlyAnchorSelection.TryResolveReFlyProvisionalAnchor` bypass and the fallback nearest-search in `FlightRecorder.UpdateAnchorDetection` / `BackgroundRecorder.UpdateBackgroundAnchorDetection`, and downgrades a Relative resume to Absolute in `RestoreTrackSectionAfterFalseAlarm`. The gate applies uniformly to top-level and parent-anchored re-fly provisionals; runtime analysis (2026-05-19 Kerbal X Probe re-fly) showed the recorder's bypass pins to the supersede target for parent-anchored provisionals identically to top-level ones, so the original `DebrisParentRecordingId` carve-out was removed before merge. The setting does NOT participate in `.pann` ConfigurationHash because it only affects `.prec` authoring, not pannotation generation, so flipping it does not invalidate cached sidecars. The setter for the toggle gates BOTH the user-flip `Anchor` log and the persistence `Record*` write on `ParsekSettingsPersistence.IsReconciled`, so KSP's GameParameters.OnLoad calling the setter to restore the field from .sfs during each scene / quicksave load (a state restore, not a user toggle) no longer emits a spurious "False->True" Anchor flip line per scene change. See `docs/dev/plans/force-absolute-refly-provisional.md` for the experiment design and known regressions (docking-mid-rewind, loop-anchored re-fly fork).
+- Re-fly Relative anchor selection now uses a narrowed-gate filter instead of the PR #889 supersede-target bypass. While a re-fly session is active, the nearest-search drops every candidate whose recording is a member of the same `RecordingTree` as the provisional, leaving out-of-tree real anchors (stations, bases, live vessels from other lineages) eligible. Net result: re-fly with no nearby real anchor authors Absolute (the PR 901 validated default), re-fly mid-docking-approach authors Relative-against-real-station (closes the PR 901 "docking-mid-rewind" regression), and re-fly inside a loop chain with a live loop anchor authors Relative-against-live-loop-anchor (closes the "loop-anchored re-fly fork" regression).
+- Removed the experimental "Force Absolute for re-fly provisional" diagnostics toggle now that the narrowed-gate default is validated.
 - Retired the Phase 5 co-bubble peer blending subsystem. The default-off toggle from earlier in v0.10.0 is gone, every ghost now renders its own standalone Absolute trajectory unconditionally, and the controlled-decoupled child parent-anchor contract introduced earlier in this release handles the same close-formation cases through `TrackSection.bodyFixedFrames` at a deeper level. Existing `.pann` sidecars regenerate on next load through the normal config-hash-drift path. Git log preserves the algorithms (`CoBubbleBlender`, `CoBubbleOverlapDetector`, `CoBubblePrimarySelector`, the `.pann` offset-trace block) before this commit's parent.
 
 ---
